+++ pydio/__init__.py
@@ -15,4 +15,4 @@ __author__ = 'Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>'
 try:
     __version__ = get_distribution(__name__).version
 except DistributionNotFound:
-    __version__ = '0.1.0rc3'  # Use 'inv tag' to update this
+    __version__ = '0.1.0rc4'  # Use 'inv tag' to update this