+++ pydio/keys.py
@@ -1,5 +1,5 @@
 # ---------------------------------------------------------------------------
-# pydio/variant.py
+# pydio/keys.py
 #
 # Copyright (C) 2021 Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>
 #
@@ -9,11 +9,22 @@
 # See LICENSE.txt for details.
 # ---------------------------------------------------------------------------
 import collections.abc
-import itertools
 from typing import Hashable
 
 
 class Variant(collections.abc.Hashable):
+    """A special form of key that can have user-defined parameters attached.
+
+    This class can be used if you need to use same key twice, but return
+    different objects depending on additional parameters given (which can be
+    accessed by object factory)
+
+    :param key:
+        The key to be wrapped
+
+    :param kwargs:
+        Additional parameters to be bound with given key
+    """
 
     def __init__(self, key: Hashable, **kwargs: Hashable):
         self._key = key
@@ -21,10 +32,12 @@ class Variant(collections.abc.Hashable):
 
     @property
     def key(self):
+        """Wrapped key."""
         return self._key
 
     @property
-    def kwargs(self):
+    def kwargs(self) -> dict:
+        """Dict with parameters given in constructor."""
         return self._kwargs
 
     def __repr__(self):
@@ -33,9 +46,7 @@ class Variant(collections.abc.Hashable):
         )
 
     def __hash__(self):
-        return hash(
-            frozenset(itertools.chain([self._key], self._kwargs.items()))
-        )
+        return hash(self._key)
 
     def __eq__(self, other):
         return isinstance(other, self.__class__) and\