+++ tests/unit/test_factories.py
@@ -1,5 +1,5 @@
 # ---------------------------------------------------------------------------
-# tests/unit/test_factory.py
+# tests/unit/test_factories.py
 #
 # Copyright (C) 2021 Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>
 #
@@ -13,24 +13,20 @@ from mockify.actions import Return
 from mockify.core import satisfied
 from mockify.mock import Mock
 
-from pydio import _factory as factory
-from pydio import base
+from pydio import _factories as factories
 
 
 class TestGeneratorFactory:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.uut = factory.GeneratorFactory(self.make_connection)
+        self.uut = factories.GeneratorFactory(self.make_connection)
         self.database = Mock('database')
 
     def make_connection(self):
         connection = self.database.connect()
         yield connection
         connection.close()
-
-    def test_generator_factory_is_not_awaitable(self):
-        assert not self.uut.is_awaitable()
 
     def test_get_instance_always_returns_same_object(self):
         connection = Mock('connection')
@@ -48,25 +44,22 @@ class TestGeneratorFactory:
         with satisfied(connection):
             self.uut.close()
 
-    def test_get_instance_returns_null_if_called_after_close(self):
+    def test_get_instance_returns_none_if_called_after_close(self):
         self.uut.close()
-        assert self.uut.get_instance() is base.NULL
+        assert self.uut.get_instance() is None
 
 
 class TestAsyncGeneratorFactory:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.uut = factory.AsyncGeneratorFactory(self.make_connection)
+        self.uut = factories.AsyncGeneratorFactory(self.make_connection)
         self.database = Mock('database')
 
     async def make_connection(self):
         connection = self.database.connect()
         yield connection
         connection.close()
-
-    def test_async_generator_factory_is_awaitable(self):
-        assert self.uut.is_awaitable()
 
     @pytest.mark.asyncio
     async def test_get_instance_always_returns_same_object(self):
@@ -89,22 +82,19 @@ class TestAsyncGeneratorFactory:
             await self.uut.close()
 
     @pytest.mark.asyncio
-    async def test_get_instance_returns_null_if_called_after_close(self):
+    async def test_get_instance_returns_none_if_called_after_close(self):
         await self.uut.close()
-        assert await self.uut.get_instance() is base.NULL
+        assert await self.uut.get_instance() is None
 
 
 class TestCoroutineFactory:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.uut = factory.CoroutineFactory(self.make_dummy_object)
+        self.uut = factories.CoroutineFactory(self.make_dummy_object)
 
     async def make_dummy_object(self):
         return [42]
-
-    def test_coroutine_factory_is_awaitable(self):
-        assert self.uut.is_awaitable()
 
     @pytest.mark.asyncio
     async def test_get_instance_always_returns_same_object(self):
@@ -114,22 +104,19 @@ class TestCoroutineFactory:
         assert first == [42]
 
     @pytest.mark.asyncio
-    async def test_get_instance_returns_null_after_close(self):
+    async def test_get_instance_returns_none_after_close(self):
         await self.uut.close()
-        assert await self.uut.get_instance() is base.NULL
+        assert await self.uut.get_instance() is None
 
 
 class TestFunctionFactory:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.uut = factory.FunctionFactory(self.make_dummy_object)
+        self.uut = factories.FunctionFactory(self.make_dummy_object)
 
     def make_dummy_object(self):
         return [42]
-
-    def test_function_factory_is_not_awaitable(self):
-        assert not self.uut.is_awaitable()
 
     def test_get_instance_always_returns_same_object(self):
         first = self.uut.get_instance()
@@ -137,24 +124,21 @@ class TestFunctionFactory:
         assert first is second
         assert first == [42]
 
-    def test_get_instance_returns_null_after_close(self):
+    def test_get_instance_returns_none_after_close(self):
         self.uut.close()
-        assert self.uut.get_instance() is base.NULL
+        assert self.uut.get_instance() is None
 
 
 class TestInstanceFactory:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.uut = factory.InstanceFactory([42])
-
-    def test_instance_factory_is_not_awaitable(self):
-        assert not self.uut.is_awaitable()
+        self.uut = factories.InstanceFactory([42])
 
     def test_get_instance_always_returns_same_object(self):
         assert self.uut.get_instance() is self.uut.get_instance()
         assert self.uut.get_instance() == [42]
 
-    def test_get_instance_returns_null_after_close(self):
+    def test_get_instance_returns_none_after_close(self):
         self.uut.close()
-        assert self.uut.get_instance() is base.NULL
+        assert self.uut.get_instance() is None