+++ tests/unit/test_keys.py
@@ -1,5 +1,5 @@
 # ---------------------------------------------------------------------------
-# tests/unit/test_variant.py
+# tests/unit/test_keys.py
 #
 # Copyright (C) 2021 Maciej Wiatrzyk <maciej.wiatrzyk@gmail.com>
 #
@@ -8,7 +8,7 @@
 #
 # See LICENSE.txt for details.
 # ---------------------------------------------------------------------------
-from pydio.variant import Variant
+from pydio.keys import Variant
 
 
 class TestVariant: