+++ CHANGELOG.rst
@@ -1,3 +1,19 @@
+0.1.0rc4 (2021-02-10)
+---------------------
+
+**Added**
+
+  * Add missing support for async context manager in
+    :class:`pydio.injector.Injector` class
+  * Add initial documentation draft with API docs
+  * Add some badges to the project
+
+**Changed**
+
+  * Improved README.rst file with simple example
+  * Redesigned :mod:`pydio.base` module
+  * Module renamed: pydio.variant -> :mod:`pydio.keys`
+
 0.1.0rc3 (2021-02-04)
 ---------------------